Update room's foundWords on a fixed interval
For now, this interval is every ten seconds. In a perfect world, each component would be able to detect a change to the room in the database and update immediately, but I haven't worked out that kind of persistent connection yet.

Also allows players to manually refresh foundWords with a button.